+++ .vscode/settings.json
@@ -92,6 +92,8 @@
         "typeinfo": "cpp",
         "valarray": "cpp",
         "variant": "cpp",
-        "regex": "cpp"
+        "regex": "cpp",
+        "forward_list": "cpp",
+        "text_encoding": "cpp"
     },
 }
