+++ engine/src/EntryPoint.h
@@ -9,7 +9,7 @@ int main(int argc, char** argv) {
     JC2D::Log::Init();
     auto app = JC2D::CreateApplication();
     JC2D_CORE_INFO("Start");
-    app->Run();
+    app->run();
     delete app;
     return 0;
 }
