+++ engine/src/core/event/event.h
@@ -1,7 +1,7 @@
 #pragma once
-#include "../../jc2dpch.h"
 
-#include "./defines.h"
+#include "../../jc2dpch.h"
+#include "../../defines.h"
 
 namespace JC2D {
     enum class EventType {
@@ -23,7 +23,7 @@ namespace JC2D {
     };
 
 #define EVENT_CLASS_TYPE(type)                                                  \
-    static EventType GetStaticType() { return EventType::##type; }              \
+    static EventType GetStaticType() { return EventType::type; }                \
     virtual EventType GetEventType() const override { return GetStaticType(); } \
     virtual const char* GetName() const override { return #type; }
 
@@ -62,7 +62,7 @@ namespace JC2D {
 
        public:
         EventDispatcher(Event& event)
-            : m_Event(event){};
+            : m_Event(event) {};
 
         template <typename T>
         bool Dispatch(EventFn<T> func) {
