+++ engine/src/core/window.h
@@ -1,7 +1,4 @@
-#include "defines.h"
-#include "jc2dpch.h"
-
-#include "./core/event/event.h"
+#include "./event/event.h"
 
 namespace JC2D {
     struct WindowProps {
@@ -19,14 +16,14 @@ namespace JC2D {
        public:
         using EventCallbackFn = std::function<void(Event&)>;
 
-        virtual ~Window() {};
+        virtual ~Window() = default;
 
         virtual void OnUpdate() = 0;
 
         virtual unsigned int GetWidth() const = 0;
         virtual unsigned int GetHeight() const = 0;
 
-        virtual void SetEventCallback(const EventCallbackFn& callback);
+        virtual void SetEventCallback(const EventCallbackFn& callback) = 0;
         virtual void SetVSync(bool enabled) = 0;
         virtual bool IsVSync() const = 0;
