+++ engine/src/jc2dpch.h
@@ -19,6 +19,3 @@
 #include "entt.hpp"
 #include "stb_image.h"
 #include "spdlog/spdlog.h"
-
-#include "logger/logger.h"
-#include "asserts.h"
