+++ engine/src/platform/linux/linux_window.h
@@ -1,4 +1,4 @@
-#include "window.h"
+#include "../../core/window.h"
 
 namespace JC2D {
     class LinuxWindow : public Window {
